ui/app.py
"""
Aplicación Streamlit para detección de objetos sospechosos
"""
import streamlit as st
import cv2
import numpy as np
from PIL import Image

# Importaciones desde src
from src.detector import detect_objects, load_model, draw_detections
from src.preprocessing import preprocess_frame
from src.logic import check_security_risk, calculate_risk_level
from src.utils import save_image, get_timestamp


def main():
    """
    Función principal de la aplicación Streamlit
    """
    st.set_page_config(
        page_title="Sistema de Detección de Seguridad",
        page_icon="🔒",
        layout="wide"
    )
    
    st.title("🔒 Sistema de Detección de Objetos Sospechosos")
    st.markdown("### Detección en tiempo real usando CNN")
    
    # Sidebar con opciones
    setup_sidebar()
    
    # Tabs principales
    tab1, tab2, tab3 = st.tabs(["📸 Detección en Imagen", "🎥 Detección en Video", "📊 Historial"])
    
    with tab1:
        image_detection_tab()
    
    with tab2:
        video_detection_tab()
    
    with tab3:
        history_tab()


def setup_sidebar():
    """
    Configura el panel lateral con opciones
    """
    st.sidebar.header("⚙️ Configuración")
    st.sidebar.slider("Umbral de Confianza", 0.0, 1.0, 0.5, 0.05)
    st.sidebar.slider("Umbral de Riesgo", 0.0, 1.0, 0.7, 0.05)
    st.sidebar.selectbox("Modelo CNN", ["YOLO", "ResNet", "MobileNet"])
    st.sidebar.checkbox("Guardar Detecciones", value=True)


def image_detection_tab():
    """
    Tab para detección en imágenes estáticas
    """
    st.header("Detección en Imagen")
    
    uploaded_file = st.file_uploader("Cargar imagen", type=['jpg', 'jpeg', 'png'])
    
    if uploaded_file is not None:
        # Mostrar imagen original
        col1, col2 = st.columns(2)
        
        with col1:
            st.subheader("Imagen Original")
            # TODO: Procesar y mostrar imagen
            pass
        
        with col2:
            st.subheader("Detecciones")
            # TODO: Mostrar resultados de detección
            pass
        
        # Mostrar análisis de riesgo
        show_risk_analysis()


def video_detection_tab():
    """
    Tab para detección en video/webcam
    """
    st.header("Detección en Video")
    
    option = st.radio("Seleccionar fuente:", ["Subir video", "Usar webcam"])
    
    if option == "Subir video":
        uploaded_video = st.file_uploader("Cargar video", type=['mp4', 'avi', 'mov'])
        # TODO: Procesar video
        pass
    else:
        if st.button("Iniciar Webcam"):
            # TODO: Implementar detección en webcam
            pass


def history_tab():
    """
    Tab para mostrar historial de detecciones
    """
    st.header("Historial de Detecciones")
    
    # TODO: Mostrar tabla con historial
    st.info("Aquí se mostrarán las detecciones guardadas")


def show_risk_analysis():
    """
    Muestra el análisis de riesgo de seguridad
    """
    st.subheader("📊 Análisis de Riesgo")
    
    # TODO: Mostrar métricas de riesgo
    col1, col2, col3 = st.columns(3)
    
    with col1:
        st.metric("Nivel de Riesgo", "Medio")
    
    with col2:
        st.metric("Objetos Detectados", "0")
    
    with col3:
        st.metric("Objetos Sospechosos", "0")


def process_image(image):
    """
    Procesa una imagen y realiza la detección
    
    Args:
        image: Imagen a procesar
    
    Returns:
        Resultados de la detección
    """
    pass


def process_video_frame(frame):
    """
    Procesa un frame de video
    
    Args:
        frame: Frame a procesar
    
    Returns:
        Frame procesado con detecciones
    """
    pass


if __name__ == "__main__":
    main()
