requirements.txt
# Deep Learning Frameworks
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# UI Framework
streamlit>=1.28.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Others
python-dotenv>=1.0.0
